Added sim cases 7 and 8 with gmsh rectangular plates with holes
